+++ src/api/url.ts
@@ -115,7 +115,7 @@ export const customFetch = async (
 
     const isAuthRoute = input.includes('/auth'); // 로그인, 회원가입 경로 확인
 
-    if ([401, 419].includes(response.status)) {
+    if ([419].includes(response.status)) {
       // 로그인 경로가 아닌 경우 리다이렉트
       handleAuthError(message, !isAuthRoute, options.toastId);
     } else if (response.status === 403) {
